+++ src/app/modules/submission/encrypt-submission/encrypt-submission.constants.ts
@@ -1,6 +1,1 @@
-/**
- * 60 seconds = 1 minute. The expiry time for presigned POST URLs.
- */
-export const PRESIGNED_ATTACHMENT_POST_EXPIRY_SECS = 60
-
 export const CHARTS_MAX_SUBMISSION_RESULTS = 1000
