+++ src/app/modules/submission/encrypt-submission/encrypt-submission.controller.ts
@@ -5,14 +5,10 @@ import { StatusCodes } from 'http-status-codes'
 import mongoose from 'mongoose'
 import Stripe from 'stripe'
 
-import { featureFlags } from '../../../../../shared/constants'
 import {
-  AttachmentPresignedPostDataMapType,
-  AttachmentSizeMapType,
   DateString,
   ErrorDto,
   FormAuthType,
-  FormResponseMode,
   Payment,
   PaymentChannel,
   PaymentType,
@@ -38,7 +34,6 @@ import { createReqMeta } from '../../../utils/request'
 import { getFormAfterPermissionChecks } from '../../auth/auth.service'
 import { ControllerHandler } from '../../core/core.types'
 import { setFormTags } from '../../datadog/datadog.utils'
-import { getFeatureFlag } from '../../feature-flags/feature-flags.service'
 import { PermissionLevel } from '../../form/admin-form/admin-form.types'
 import { MyInfoService } from '../../myinfo/myinfo.service'
 import { extractMyInfoLoginJwt } from '../../myinfo/myinfo.util'
@@ -48,7 +43,9 @@ import { getPopulatedUserById } from '../../user/user.service'
 import * as VerifiedContentService from '../../verified-content/verified-content.service'
 import * as EncryptSubmissionMiddleware from '../encrypt-submission/encrypt-submission.middleware'
 import * as ReceiverMiddleware from '../receiver/receiver.middleware'
-import { fileSizeLimit, fileSizeLimitBytes } from '../submission.utils'
+import { SubmissionFailedError } from '../submission.errors'
+import { uploadAttachments } from '../submission.service'
+import { mapRouteError } from '../submission.utils'
 import { reportSubmissionResponseTime } from '../submissions.statsd-client'
 
 import {
@@ -57,15 +54,9 @@ import {
   ensureValidCaptcha,
 } from './encrypt-submission.ensures'
 import {
-  FeatureDisabledError,
-  SubmissionFailedError,
-} from './encrypt-submission.errors'
-import {
   checkFormIsEncryptMode,
   getAllEncryptedSubmissionData,
-  getQuarantinePresignedPostData,
   performEncryptPostSubmissionActions,
-  uploadAttachments,
 } from './encrypt-submission.service'
 import {
   SubmitEncryptModeFormHandlerRequest,
@@ -75,7 +66,6 @@ import {
   getPaymentAmount,
   getPaymentIntentDescription,
   getStripePaymentMethod,
-  mapRouteError,
 } from './encrypt-submission.utils'
 
 const logger = createLoggerWithLabel(module)
@@ -730,114 +720,4 @@ export const handleGetAllEncryptedResponses = [
       .and('startDate', 'endDate'),
   }),
   _getAllEncryptedResponse,
-] as ControllerHandler[]
-
-/**
- * Handler for POST /:formId/submissions/storage/get-s3-presigned-post-data
- * Used by handleGetS3PresignedPostData after joi validation
- * @returns 200 with array of presigned post data
- * @returns 400 if ids are invalid or total file size exceeds 20MB
- * @returns 500 if presigned post data cannot be retrieved or any other errors occur
- * Exported for testing
- */
-export const getS3PresignedPostData: ControllerHandler<
-  unknown,
-  AttachmentPresignedPostDataMapType[] | ErrorDto,
-  AttachmentSizeMapType[]
-> = async (req, res) => {
-  const logMeta = {
-    action: 'getS3PresignedPostData',
-    ...createReqMeta(req),
-  }
-
-  return getFeatureFlag(featureFlags.encryptionBoundaryShiftVirusScanner)
-    .map((virusScannerEnabled) => {
-      if (!virusScannerEnabled) {
-        logger.warn({
-          message: 'Virus scanning has not been enabled.',
-          meta: logMeta,
-        })
-
-        const { statusCode, errorMessage } = mapRouteError(
-          new FeatureDisabledError(),
-        )
-        return res.status(statusCode).send({
-          message: errorMessage,
-        })
-      }
-
-      return getQuarantinePresignedPostData(req.body)
-        .map((presignedUrls) => {
-          logger.info({
-            message: 'Successfully retrieved quarantine presigned post data.',
-            meta: logMeta,
-          })
-          return res.status(StatusCodes.OK).send(presignedUrls)
-        })
-        .mapErr((error) => {
-          logger.error({
-            message: 'Failure getting quarantine presigned post data.',
-            meta: logMeta,
-            error,
-          })
-          const { statusCode, errorMessage } = mapRouteError(error)
-          return res.status(statusCode).send({
-            message: errorMessage,
-          })
-        })
-    })
-    .mapErr((error) => {
-      logger.error({
-        message: 'Error retrieving feature flags.',
-        meta: logMeta,
-        error,
-      })
-      const { statusCode, errorMessage } = mapRouteError(error)
-      return res.status(statusCode).send({
-        message: errorMessage,
-      })
-    })
-}
-
-/**
- * Custom validation function for Joi to validate that the sum of 'size' in the array of objects
- * is less than or equal to total file size limit (20MB).
- */
-const validateFileSizeSum = (
-  value: { size: number }[],
-  helpers: { error: (arg0: string) => null },
-) => {
-  const sum = value.reduce((acc, curr) => acc + curr.size, 0)
-
-  if (sum <= fileSizeLimitBytes(FormResponseMode.Encrypt)) {
-    return value // Return the validated value if the sum of 'size' is less than or equal to limit
-  } else {
-    return helpers.error('size.limit') // Return an error if the sum of 'size' is greater than limit
-  }
-}
-
-// Handler for POST /:formId/submissions/storage/get-s3-presigned-post-data
-export const handleGetS3PresignedPostData = [
-  celebrate({
-    [Segments.BODY]: Joi.array()
-      .items(
-        Joi.object().keys({
-          id: Joi.string()
-            .regex(/^[0-9a-fA-F]{24}$/) // IDs should be MongoDB ObjectIDs
-            .required(),
-          size: Joi.number()
-            .max(fileSizeLimitBytes(FormResponseMode.Encrypt)) // Max attachment size is 20MB
-            .required(),
-        }),
-      )
-      .unique('id') // IDs of each array item should be unique
-      .custom(validateFileSizeSum, 'Custom validation for total file size') // Custom validation to check for total file size
-      .messages({
-        'size.limit': `Total file size exceeds ${fileSizeLimit(
-          FormResponseMode.Encrypt,
-        )}MB`, // Custom error message for total file size
-        'array.unique': 'Duplicate id(s) found', // Custom error message for duplicate IDs
-      }),
-  }),
-  getS3PresignedPostData,
 ] as ControllerHandler[]
