+++ src/app/modules/submission/encrypt-submission/encrypt-submission.types.ts
@@ -1,10 +1,7 @@
-import { StatusCodes } from 'http-status-codes'
-
 import {
   SubmissionErrorDto,
   SubmissionResponseDto,
 } from '../../../../../shared/types'
-import { hasProp } from '../../../../../shared/utils/has-prop'
 import { IPopulatedEncryptedForm } from '../../../../types'
 import {
   EncryptFormLoadedDto,
@@ -14,10 +11,6 @@ import {
   ParsedStorageModeSubmissionBody,
 } from '../../../../types/api'
 import { ControllerHandler } from '../../core/core.types'
-
-import { ParseVirusScannerLambdaPayloadError } from './encrypt-submission.errors'
-
-export type AttachmentMetadata = Map<string, string>
 
 export type SaveEncryptSubmissionParams = {
   form: IPopulatedEncryptedForm
@@ -76,72 +69,3 @@ export type SubmitEncryptModeFormHandlerType = ControllerHandler<
 
 export type SubmitEncryptModeFormHandlerRequest =
   Parameters<SubmitEncryptModeFormHandlerType>[0] & { formsg: FormCompleteDto }
-
-export type ParseVirusScannerLambdaPayloadBeforeBodyIsParsed = {
-  statusCode: number
-  body: string
-}
-
-export type ParseVirusScannerLambdaPayloadOkBody = {
-  cleanFileKey: string
-  destinationVersionId: string
-}
-
-export type ParseVirusScannerLambdaPayloadOkType = {
-  statusCode: StatusCodes.OK
-  body: ParseVirusScannerLambdaPayloadOkBody
-}
-
-export type ParseVirusScannerLambdaPayloadErrBody = {
-  message: string
-}
-
-export type ParseVirusScannerLambdaPayloadErrType =
-  | {
-      statusCode: number // custom status codes might be sent by the lambda
-      body: ParseVirusScannerLambdaPayloadErrBody
-    }
-  | ParseVirusScannerLambdaPayloadError
-
-// Helper function to check if the payload is of the expected structure
-export const payloadIsExpectedStructure = (
-  parsedPayload: unknown,
-): false | ParseVirusScannerLambdaPayloadBeforeBodyIsParsed => {
-  return (
-    typeof parsedPayload === 'object' &&
-    !!parsedPayload &&
-    hasProp(parsedPayload, 'statusCode') &&
-    typeof parsedPayload.statusCode === 'number' &&
-    hasProp(parsedPayload, 'body') &&
-    typeof parsedPayload.body === 'string' &&
-    (parsedPayload as ParseVirusScannerLambdaPayloadBeforeBodyIsParsed)
-  )
-}
-
-// Helper function to check if the body is of the expected structure for OK status code
-export const bodyIsExpectedOkStructure = (
-  parsedBody: unknown,
-): false | ParseVirusScannerLambdaPayloadOkBody => {
-  return (
-    typeof parsedBody === 'object' &&
-    !!parsedBody &&
-    hasProp(parsedBody, 'cleanFileKey') &&
-    typeof parsedBody.cleanFileKey === 'string' &&
-    hasProp(parsedBody, 'destinationVersionId') &&
-    typeof parsedBody.destinationVersionId === 'string' &&
-    (parsedBody as ParseVirusScannerLambdaPayloadOkBody)
-  )
-}
-
-// Helper function to check if the body is of the expected structure for non-OK status code
-export const bodyIsExpectedErrStructure = (
-  parsedBody: unknown,
-): false | ParseVirusScannerLambdaPayloadErrBody => {
-  return (
-    typeof parsedBody === 'object' &&
-    !!parsedBody &&
-    hasProp(parsedBody, 'message') &&
-    typeof parsedBody.message === 'string' &&
-    (parsedBody as ParseVirusScannerLambdaPayloadErrBody)
-  )
-}
