+++ src/app/modules/submission/encrypt-submission/encrypt-submission.utils.ts
@@ -1,4 +1,3 @@
-import { StatusCodes } from 'http-status-codes'
 import moment from 'moment-timezone'
 import Stripe from 'stripe'
 
@@ -19,286 +18,15 @@ import {
   IEncryptedSubmissionSchema,
   IPopulatedEncryptedForm,
   ISubmissionSchema,
-  MapRouteErrors,
   StorageModeSubmissionData,
 } from '../../../../types'
 import {
   EncryptFormFieldResponse,
   ParsedClearFormFieldResponse,
 } from '../../../../types/api'
-import { MapRouteError } from '../../../../types/routing'
-import { createLoggerWithLabel } from '../../../config/logger'
-import { MalformedVerifiedContentError } from '../../../modules/verified-content/verified-content.errors'
-import {
-  CaptchaConnectionError,
-  MissingCaptchaError,
-  VerifyCaptchaError,
-} from '../../../services/captcha/captcha.errors'
-import {
-  MissingTurnstileError,
-  TurnstileConnectionError,
-  VerifyTurnstileError,
-} from '../../../services/turnstile/turnstile.errors'
-import { CreatePresignedPostError } from '../../../utils/aws-s3'
-import { genericMapRouteErrorTransform } from '../../../utils/error'
-import {
-  AttachmentUploadError,
-  DatabaseConflictError,
-  DatabaseError,
-  DatabasePayloadSizeError,
-  DatabaseValidationError,
-  EmptyErrorFieldError,
-  MalformedParametersError,
-} from '../../core/core.errors'
-import {
-  ForbiddenFormError,
-  FormDeletedError,
-  FormNotFoundError,
-  PrivateFormError,
-} from '../../form/form.errors'
-import {
-  MyInfoCookieStateError,
-  MyInfoHashDidNotMatchError,
-  MyInfoHashingError,
-  MyInfoInvalidLoginCookieError,
-  MyInfoMissingHashError,
-  MyInfoMissingLoginCookieError,
-} from '../../myinfo/myinfo.errors'
 import { MyInfoKey } from '../../myinfo/myinfo.types'
-import { PaymentNotFoundError } from '../../payments/payments.errors'
-import {
-  SgidInvalidJwtError,
-  SgidMissingJwtError,
-  SgidVerifyJwtError,
-} from '../../sgid/sgid.errors'
-import {
-  CreateRedirectUrlError,
-  InvalidJwtError,
-  MissingJwtError,
-  VerifyJwtError,
-} from '../../spcp/spcp.errors'
-import { MissingUserError } from '../../user/user.errors'
-import {
-  AttachmentTooLargeError,
-  ConflictError,
-  InvalidEncodingError,
-  InvalidFileExtensionError,
-  ProcessingError,
-  ResponseModeError,
-  SubmissionNotFoundError,
-  ValidateFieldError,
-} from '../submission.errors'
 import { ProcessedFieldResponse } from '../submission.types'
 import { getAnswersForChild, getMyInfoPrefix } from '../submission.utils'
-
-import {
-  AttachmentSizeLimitExceededError,
-  DownloadCleanFileFailedError,
-  FeatureDisabledError,
-  InvalidFieldIdError,
-  InvalidFileKeyError,
-  MaliciousFileDetectedError,
-  SubmissionFailedError,
-  VirusScanFailedError,
-} from './encrypt-submission.errors'
-
-const logger = createLoggerWithLabel(module)
-
-/**
- * Handler to map ApplicationErrors to their correct status code and error
- * messages.
- * @param error The error to retrieve the status codes and error messages
- */
-const errorMapper: MapRouteError = (
-  error,
-  coreErrorMessage = 'Sorry, something went wrong. Please try again.',
-) => {
-  switch (error.constructor) {
-    case AttachmentUploadError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'Could not upload attachments for submission. For assistance, please contact the person who asked you to fill in this form.',
-      }
-    case CreateRedirectUrlError:
-      return {
-        statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
-        errorMessage: coreErrorMessage,
-      }
-    case SgidMissingJwtError:
-    case SgidVerifyJwtError:
-    case SgidInvalidJwtError:
-    case MissingJwtError:
-    case VerifyJwtError:
-    case InvalidJwtError:
-    case MyInfoMissingLoginCookieError:
-    case MyInfoCookieStateError:
-    case MyInfoInvalidLoginCookieError:
-    case MalformedVerifiedContentError:
-      return {
-        statusCode: StatusCodes.UNAUTHORIZED,
-        errorMessage:
-          'Something went wrong with your login. Please try logging in and submitting again.',
-      }
-    case MyInfoMissingHashError:
-      return {
-        statusCode: StatusCodes.GONE,
-        errorMessage:
-          'MyInfo verification expired, please refresh and try again.',
-      }
-    case MyInfoHashDidNotMatchError:
-      return {
-        statusCode: StatusCodes.UNAUTHORIZED,
-        errorMessage: 'MyInfo verification failed.',
-      }
-    case MyInfoHashingError:
-      return {
-        statusCode: StatusCodes.SERVICE_UNAVAILABLE,
-        errorMessage:
-          'MyInfo verification unavailable, please try again later.',
-      }
-    case MissingUserError:
-      return {
-        statusCode: StatusCodes.UNPROCESSABLE_ENTITY,
-        errorMessage: error.message,
-      }
-    case FormNotFoundError:
-      return {
-        statusCode: StatusCodes.NOT_FOUND,
-        errorMessage: error.message,
-      }
-    case ResponseModeError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage: error.message,
-      }
-    case FeatureDisabledError:
-    case ForbiddenFormError:
-      return {
-        statusCode: StatusCodes.FORBIDDEN,
-        errorMessage: error.message,
-      }
-    case FormDeletedError:
-      return {
-        statusCode: StatusCodes.GONE,
-        errorMessage: error.message,
-      }
-    case PrivateFormError:
-      return {
-        statusCode: StatusCodes.NOT_FOUND,
-        errorMessage:
-          'This form has been taken down. For assistance, please contact the person who asked you to fill in this form.',
-      }
-    case CaptchaConnectionError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'Could not verify captcha. Please submit again in a few minutes.',
-      }
-    case VerifyCaptchaError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage: 'Captcha was incorrect. Please submit again.',
-      }
-    case MissingCaptchaError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage: 'Captcha was missing. Please refresh and submit again.',
-      }
-    case TurnstileConnectionError:
-      return {
-        statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
-        errorMessage:
-          'Error connecting to Turnstile server . Please submit again in a few minutes.',
-      }
-    case VerifyTurnstileError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'Incorrect Turnstile parameters. Please refresh and submit again.',
-      }
-    case MissingTurnstileError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'Missing Turnstile challenge. Please refresh and submit again.',
-      }
-    case MalformedParametersError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage: error.message,
-      }
-    case SubmissionNotFoundError:
-      return {
-        statusCode: StatusCodes.NOT_FOUND,
-        errorMessage: error.message,
-      }
-    case InvalidEncodingError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'Invalid data was found. Please check your responses and submit again.',
-      }
-    case DatabasePayloadSizeError:
-      return {
-        statusCode: StatusCodes.REQUEST_TOO_LONG,
-        errorMessage:
-          'Submission too large to be saved. Please reduce the size of your submission and try again.',
-      }
-    case ValidateFieldError:
-    case DatabaseValidationError:
-    case InvalidFileExtensionError:
-    case AttachmentTooLargeError:
-    case ProcessingError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage:
-          'There is something wrong with your form submission. Please check your responses and try again. If the problem persists, please refresh the page.',
-      }
-    case DatabaseConflictError:
-    case ConflictError:
-      return {
-        statusCode: StatusCodes.CONFLICT,
-        errorMessage:
-          'The form has been updated. Please refresh and submit again.',
-      }
-    case PaymentNotFoundError:
-    case CreatePresignedPostError:
-    case DatabaseError:
-    case EmptyErrorFieldError:
-    case VirusScanFailedError:
-    case DownloadCleanFileFailedError:
-      return {
-        statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
-        errorMessage: error.message,
-      }
-    case SubmissionFailedError:
-    case InvalidFieldIdError:
-    case AttachmentSizeLimitExceededError:
-    case InvalidFileKeyError:
-    case MaliciousFileDetectedError:
-      return {
-        statusCode: StatusCodes.BAD_REQUEST,
-        errorMessage: error.message,
-      }
-    default:
-      logger.error({
-        message: 'Unknown route error observed',
-        meta: {
-          action: 'mapRouteError',
-        },
-        error,
-      })
-
-      return {
-        statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
-        errorMessage: 'Something went wrong. Please try again.',
-      }
-  }
-}
-
-export const mapRouteError: MapRouteErrors =
-  genericMapRouteErrorTransform(errorMapper)
 
 /**
  * Typeguard to check if given submission is an encrypt mode submission.
