+++ src/app/modules/submission/submission.constants.ts
@@ -29,3 +29,8 @@ export const sharedSubmissionParams = {
     numVisibleFields: Joi.number(),
   }),
 }
+
+/**
+ * 60 seconds = 1 minute. The expiry time for presigned POST URLs.
+ */
+export const PRESIGNED_ATTACHMENT_POST_EXPIRY_SECS = 60
