+++ src/app/modules/submission/submission.errors.ts
@@ -96,3 +96,75 @@ export class InvalidFileExtensionError extends ApplicationError {
     super(message)
   }
 }
+
+export class SubmissionFailedError extends ApplicationError {
+  constructor(
+    message = 'The form submission could not be processed. Please try again.',
+  ) {
+    super(message)
+  }
+}
+
+export class InvalidFieldIdError extends ApplicationError {
+  constructor(
+    message = 'Invalid field id. Field id should be a valid MongoDB ObjectId.',
+  ) {
+    super(message)
+  }
+}
+
+export class AttachmentSizeLimitExceededError extends ApplicationError {
+  constructor(
+    message = `Total attachment size exceeds maximum file size limit. Please reduce your total attachment size and try again.`,
+  ) {
+    super(message)
+  }
+}
+
+export class FeatureDisabledError extends ApplicationError {
+  constructor(message = 'This feature is disabled.') {
+    super(message)
+  }
+}
+
+export class InvalidFileKeyError extends ApplicationError {
+  constructor(message = 'Invalid file key. File keys should be valid UUIDs.') {
+    super(message)
+  }
+}
+
+export class VirusScanFailedError extends ApplicationError {
+  constructor(message = 'Virus scan failed. Please try again.') {
+    super(message)
+  }
+}
+
+export class JsonParseFailedError extends ApplicationError {
+  constructor(message = 'JSON parsing failed. Please try again.') {
+    super(message)
+  }
+}
+
+export class DownloadCleanFileFailedError extends ApplicationError {
+  constructor(
+    message = 'Attempt to download clean file failed. Please try again.',
+  ) {
+    super(message)
+  }
+}
+
+export class ParseVirusScannerLambdaPayloadError extends ApplicationError {
+  constructor(message = 'Unexpected payload from virus scanning lambda.') {
+    super(message)
+  }
+}
+
+export class MaliciousFileDetectedError extends ApplicationError {
+  constructor(filename?: string) {
+    super(
+      `Your ${
+        filename ? `file "${filename}"` : 'attachments(s)'
+      } has failed our virus scan. Try to create and upload it again.`,
+    )
+  }
+}
